Add PROXY protocol CRC32C TLV coverage
